KFSMI-5507 - Inserted system parameter lookup to insert default value into object code o implement the default Object Code from system parameter is the following:
Create a new public service interface method in OrganizationReversionService , called getOrganizationReversionDetaiFromSystemParameters()
Implement the new service method in OrganizationReversionServiceImpl.  This method uses the  system parameter service to fetch the default Object Code.  In OrganizationReversionServiceImpl, I defined a setParmeterService method and inserted into the spring-coa.xml  a reference to parameter service for the organizationReversionService bean.
In OrganizationReversionPreRules.doCustomPreRules.doCustomPreRules(), I inserted code to call a new function which checks if the "carry forward by object code" indicator is set. If set then it uses Spring.context.getBean() to get  the new Service method created in step 1 to fetch the default Object Code and insert  it into the Organization Reversion Object Code field.
All this happens before it gets to validation so the validation process  correctly validates it.